Moved modules plugin into index.ts
