+++ packages/vite-plugin-cloudflare/src/index.ts
@@ -2,6 +2,7 @@ import assert from "node:assert";
 import * as fs from "node:fs";
 import * as path from "node:path";
 import { createMiddleware } from "@hattip/adapter-node";
+import MagicString from "magic-string";
 import { Miniflare } from "miniflare";
 import * as vite from "vite";
 import {
@@ -14,22 +15,19 @@ import {
 	getDevMiniflareOptions,
 	getPreviewMiniflareOptions,
 } from "./miniflare-options";
-import { modulesPlugin } from "./modules-plugin";
 import {
 	getNodeCompatAliases,
 	injectGlobalCode,
 	resolveNodeCompatId,
 } from "./node-js-compat";
 import { resolvePluginConfig } from "./plugin-config";
+import { MODULE_PATTERN } from "./shared";
 import { getOutputDirectory, toMiniflareRequest } from "./utils";
 import { handleWebSocket } from "./websockets";
 import { getWarningForWorkersConfigs } from "./workers-configs";
+import type { ModuleType } from "./constants";
 import type { PluginConfig, ResolvedPluginConfig } from "./plugin-config";
 import type { Unstable_RawConfig } from "wrangler";
-
-export interface ResolvedPluginConfigContainer {
-	resolvedPluginConfig: ResolvedPluginConfig;
-}
 
 /**
  * Vite plugin that enables a full-featured integration between Vite and the Cloudflare Workers runtime.
@@ -46,287 +44,381 @@ export function cloudflare(pluginConfig: PluginConfig = {}): vite.Plugin[] {
 	// this flag is used to show the workers configs warning only once
 	let workersConfigsWarningShown = false;
 
-	// This is temporary pending a better solution when we refactor the plugin into multiple plugins
-	const resolvedPluginConfigContainer = {} as ResolvedPluginConfigContainer;
-
-	const plugin: vite.Plugin = {
-		name: "vite-plugin-cloudflare",
-		config(userConfig, env) {
-			if (env.isPreview) {
-				return { appType: "custom" };
-			}
-
-			resolvedPluginConfig = resolvePluginConfig(pluginConfig, userConfig, env);
-			resolvedPluginConfigContainer.resolvedPluginConfig = resolvedPluginConfig;
-
-			if (!workersConfigsWarningShown) {
-				workersConfigsWarningShown = true;
-				const workersConfigsWarning = getWarningForWorkersConfigs(
-					resolvedPluginConfig.rawConfigs
-				);
-				if (workersConfigsWarning) {
-					console.warn(workersConfigsWarning);
+	return [
+		{
+			name: "vite-plugin-cloudflare",
+			config(userConfig, env) {
+				if (env.isPreview) {
+					return { appType: "custom" };
 				}
-			}
 
-			return {
-				appType: "custom",
-				resolve: {
-					alias: getNodeCompatAliases(),
-				},
-				environments:
-					resolvedPluginConfig.type === "workers"
-						? {
-								...Object.fromEntries(
-									Object.entries(resolvedPluginConfig.workers).map(
-										([environmentName, workerConfig]) => {
-											return [
-												environmentName,
-												createCloudflareEnvironmentOptions(
-													workerConfig,
-													userConfig,
-													environmentName
-												),
-											];
-										}
-									)
-								),
-								client: {
-									build: {
-										outDir: getOutputDirectory(userConfig, "client"),
-									},
-								},
-							}
-						: undefined,
-				builder: {
-					async buildApp(builder) {
-						const clientEnvironment = builder.environments.client;
-						const defaultHtmlPath = path.resolve(
-							builder.config.root,
-							"index.html"
-						);
-
-						if (
-							clientEnvironment &&
-							(clientEnvironment.config.build.rollupOptions.input ||
-								fs.existsSync(defaultHtmlPath))
-						) {
-							await builder.build(clientEnvironment);
-						}
-
-						if (resolvedPluginConfig.type === "workers") {
-							const workerEnvironments = Object.keys(
-								resolvedPluginConfig.workers
-							).map((environmentName) => {
-								const environment = builder.environments[environmentName];
-
-								assert(environment, `${environmentName} environment not found`);
-
-								return environment;
-							});
-
-							await Promise.all(
-								workerEnvironments.map((environment) =>
-									builder.build(environment)
-								)
-							);
-						}
-
-						writeDeployConfig(resolvedPluginConfig, resolvedViteConfig);
-					},
-				},
-			};
-		},
-		configResolved(config) {
-			resolvedViteConfig = config;
-		},
-		async resolveId(source) {
-			if (resolvedPluginConfig.type === "assets-only") {
-				return;
-			}
-
-			const workerConfig = resolvedPluginConfig.workers[this.environment.name];
-			if (!workerConfig) {
-				return;
-			}
-
-			return resolveNodeCompatId(this.environment, workerConfig, source);
-		},
-		async transform(code, id) {
-			if (resolvedPluginConfig.type === "assets-only") {
-				return;
-			}
-
-			const workerConfig = resolvedPluginConfig.workers[this.environment.name];
-
-			if (!workerConfig) {
-				return;
-			}
-
-			const resolvedId = await this.resolve(workerConfig.main);
-
-			if (id === resolvedId?.id) {
-				return injectGlobalCode(id, code, workerConfig);
-			}
-		},
-		generateBundle(_, bundle) {
-			let config: Unstable_RawConfig | undefined;
-
-			if (resolvedPluginConfig.type === "workers") {
-				const workerConfig =
-					resolvedPluginConfig.workers[this.environment.name];
-
-				const entryChunk = Object.entries(bundle).find(
-					([_, chunk]) => chunk.type === "chunk" && chunk.isEntry
+				resolvedPluginConfig = resolvePluginConfig(
+					pluginConfig,
+					userConfig,
+					env
 				);
 
-				if (!workerConfig || !entryChunk) {
+				if (!workersConfigsWarningShown) {
+					workersConfigsWarningShown = true;
+					const workersConfigsWarning = getWarningForWorkersConfigs(
+						resolvedPluginConfig.rawConfigs
+					);
+					if (workersConfigsWarning) {
+						console.warn(workersConfigsWarning);
+					}
+				}
+
+				return {
+					appType: "custom",
+					resolve: {
+						alias: getNodeCompatAliases(),
+					},
+					environments:
+						resolvedPluginConfig.type === "workers"
+							? {
+									...Object.fromEntries(
+										Object.entries(resolvedPluginConfig.workers).map(
+											([environmentName, workerConfig]) => {
+												return [
+													environmentName,
+													createCloudflareEnvironmentOptions(
+														workerConfig,
+														userConfig,
+														environmentName
+													),
+												];
+											}
+										)
+									),
+									client: {
+										build: {
+											outDir: getOutputDirectory(userConfig, "client"),
+										},
+									},
+								}
+							: undefined,
+					builder: {
+						async buildApp(builder) {
+							const clientEnvironment = builder.environments.client;
+							const defaultHtmlPath = path.resolve(
+								builder.config.root,
+								"index.html"
+							);
+
+							if (
+								clientEnvironment &&
+								(clientEnvironment.config.build.rollupOptions.input ||
+									fs.existsSync(defaultHtmlPath))
+							) {
+								await builder.build(clientEnvironment);
+							}
+
+							if (resolvedPluginConfig.type === "workers") {
+								const workerEnvironments = Object.keys(
+									resolvedPluginConfig.workers
+								).map((environmentName) => {
+									const environment = builder.environments[environmentName];
+
+									assert(
+										environment,
+										`${environmentName} environment not found`
+									);
+
+									return environment;
+								});
+
+								await Promise.all(
+									workerEnvironments.map((environment) =>
+										builder.build(environment)
+									)
+								);
+							}
+
+							writeDeployConfig(resolvedPluginConfig, resolvedViteConfig);
+						},
+					},
+				};
+			},
+			configResolved(config) {
+				resolvedViteConfig = config;
+			},
+			async resolveId(source) {
+				if (resolvedPluginConfig.type === "assets-only") {
 					return;
 				}
 
-				workerConfig.main = entryChunk[0];
-
-				const isEntryWorker =
-					this.environment.name ===
-					resolvedPluginConfig.entryWorkerEnvironmentName;
-
-				if (isEntryWorker && workerConfig.assets) {
-					const workerOutputDirectory = this.environment.config.build.outDir;
-					const clientOutputDirectory =
-						resolvedViteConfig.environments.client?.build.outDir;
-
-					assert(
-						clientOutputDirectory,
-						"Unexpected error: client output directory is undefined"
-					);
-
-					workerConfig.assets.directory = path.relative(
-						path.resolve(resolvedViteConfig.root, workerOutputDirectory),
-						path.resolve(resolvedViteConfig.root, clientOutputDirectory)
-					);
+				const workerConfig =
+					resolvedPluginConfig.workers[this.environment.name];
+				if (!workerConfig) {
+					return;
 				}
 
-				config = workerConfig;
+				return resolveNodeCompatId(this.environment, workerConfig, source);
+			},
+			async transform(code, id) {
+				if (resolvedPluginConfig.type === "assets-only") {
+					return;
+				}
 
-				if (workerConfig.configPath) {
-					const dotDevDotVarsContent = getDotDevDotVarsContent(
-						workerConfig.configPath,
-						resolvedPluginConfig.cloudflareEnv
+				const workerConfig =
+					resolvedPluginConfig.workers[this.environment.name];
+
+				if (!workerConfig) {
+					return;
+				}
+
+				const resolvedId = await this.resolve(workerConfig.main);
+
+				if (id === resolvedId?.id) {
+					return injectGlobalCode(id, code, workerConfig);
+				}
+			},
+			generateBundle(_, bundle) {
+				let config: Unstable_RawConfig | undefined;
+
+				if (resolvedPluginConfig.type === "workers") {
+					const workerConfig =
+						resolvedPluginConfig.workers[this.environment.name];
+
+					const entryChunk = Object.entries(bundle).find(
+						([_, chunk]) => chunk.type === "chunk" && chunk.isEntry
 					);
-					// Save a .dev.vars file to the worker's build output directory
-					// when it exists so that it will be then detected by `vite preview`
-					if (dotDevDotVarsContent) {
-						this.emitFile({
-							type: "asset",
-							fileName: ".dev.vars",
-							source: dotDevDotVarsContent,
-						});
+
+					if (!workerConfig || !entryChunk) {
+						return;
 					}
+
+					workerConfig.main = entryChunk[0];
+
+					const isEntryWorker =
+						this.environment.name ===
+						resolvedPluginConfig.entryWorkerEnvironmentName;
+
+					if (isEntryWorker && workerConfig.assets) {
+						const workerOutputDirectory = this.environment.config.build.outDir;
+						const clientOutputDirectory =
+							resolvedViteConfig.environments.client?.build.outDir;
+
+						assert(
+							clientOutputDirectory,
+							"Unexpected error: client output directory is undefined"
+						);
+
+						workerConfig.assets.directory = path.relative(
+							path.resolve(resolvedViteConfig.root, workerOutputDirectory),
+							path.resolve(resolvedViteConfig.root, clientOutputDirectory)
+						);
+					}
+
+					config = workerConfig;
+
+					if (workerConfig.configPath) {
+						const dotDevDotVarsContent = getDotDevDotVarsContent(
+							workerConfig.configPath,
+							resolvedPluginConfig.cloudflareEnv
+						);
+						// Save a .dev.vars file to the worker's build output directory
+						// when it exists so that it will be then detected by `vite preview`
+						if (dotDevDotVarsContent) {
+							this.emitFile({
+								type: "asset",
+								fileName: ".dev.vars",
+								source: dotDevDotVarsContent,
+							});
+						}
+					}
+				} else if (this.environment.name === "client") {
+					const assetsOnlyConfig = resolvedPluginConfig.config;
+
+					assetsOnlyConfig.assets.directory = ".";
+
+					const filesToAssetsIgnore = ["wrangler.json", ".dev.vars"];
+
+					this.emitFile({
+						type: "asset",
+						fileName: ".assetsignore",
+						source: `${filesToAssetsIgnore.join("\n")}\n`,
+					});
+
+					config = assetsOnlyConfig;
 				}
-			} else if (this.environment.name === "client") {
-				const assetsOnlyConfig = resolvedPluginConfig.config;
 
-				assetsOnlyConfig.assets.directory = ".";
+				if (!config) {
+					return;
+				}
 
-				const filesToAssetsIgnore = ["wrangler.json", ".dev.vars"];
+				config.no_bundle = true;
+				config.rules = [{ type: "ESModule", globs: ["**/*.js"] }];
+				// Set to `undefined` if it's an empty object so that the user doesn't see a warning about using `unsafe` fields when deploying their Worker.
+				if (config.unsafe && Object.keys(config.unsafe).length === 0) {
+					config.unsafe = undefined;
+				}
 
 				this.emitFile({
 					type: "asset",
-					fileName: ".assetsignore",
-					source: `${filesToAssetsIgnore.join("\n")}\n`,
+					fileName: "wrangler.json",
+					source: JSON.stringify(config),
+				});
+			},
+			handleHotUpdate(options) {
+				if (resolvedPluginConfig.configPaths.has(options.file)) {
+					options.server.restart();
+				}
+			},
+			async buildEnd() {
+				if (miniflare) {
+					await miniflare.dispose();
+					miniflare = undefined;
+				}
+			},
+			async configureServer(viteDevServer) {
+				assert(
+					viteDevServer.httpServer,
+					"Unexpected error: No Vite HTTP server"
+				);
+
+				miniflare = new Miniflare(
+					getDevMiniflareOptions(resolvedPluginConfig, viteDevServer)
+				);
+
+				await initRunners(resolvedPluginConfig, viteDevServer, miniflare);
+				const entryWorker = await getDevEntryWorker(
+					resolvedPluginConfig,
+					miniflare
+				);
+
+				const middleware = createMiddleware(({ request }) => {
+					return entryWorker.fetch(toMiniflareRequest(request), {
+						redirect: "manual",
+					}) as any;
 				});
 
-				config = assetsOnlyConfig;
-			}
+				handleWebSocket(
+					viteDevServer.httpServer,
+					entryWorker.fetch,
+					viteDevServer.config.logger
+				);
 
-			if (!config) {
-				return;
-			}
+				return () => {
+					viteDevServer.middlewares.use((req, res, next) => {
+						middleware(req, res, next);
+					});
+				};
+			},
+			configurePreviewServer(vitePreviewServer) {
+				const miniflare = new Miniflare(
+					getPreviewMiniflareOptions(
+						vitePreviewServer,
+						pluginConfig.persistState ?? true
+					)
+				);
 
-			config.no_bundle = true;
-			config.rules = [{ type: "ESModule", globs: ["**/*.js"] }];
-			// Set to `undefined` if it's an empty object so that the user doesn't see a warning about using `unsafe` fields when deploying their Worker.
-			if (config.unsafe && Object.keys(config.unsafe).length === 0) {
-				config.unsafe = undefined;
-			}
-
-			this.emitFile({
-				type: "asset",
-				fileName: "wrangler.json",
-				source: JSON.stringify(config),
-			});
-		},
-		handleHotUpdate(options) {
-			if (resolvedPluginConfig.configPaths.has(options.file)) {
-				options.server.restart();
-			}
-		},
-		async buildEnd() {
-			if (miniflare) {
-				await miniflare.dispose();
-				miniflare = undefined;
-			}
-		},
-		async configureServer(viteDevServer) {
-			assert(viteDevServer.httpServer, "Unexpected error: No Vite HTTP server");
-
-			miniflare = new Miniflare(
-				getDevMiniflareOptions(resolvedPluginConfig, viteDevServer)
-			);
-
-			await initRunners(resolvedPluginConfig, viteDevServer, miniflare);
-			const entryWorker = await getDevEntryWorker(
-				resolvedPluginConfig,
-				miniflare
-			);
-
-			const middleware = createMiddleware(({ request }) => {
-				return entryWorker.fetch(toMiniflareRequest(request), {
-					redirect: "manual",
-				}) as any;
-			});
-
-			handleWebSocket(
-				viteDevServer.httpServer,
-				entryWorker.fetch,
-				viteDevServer.config.logger
-			);
-
-			return () => {
-				viteDevServer.middlewares.use((req, res, next) => {
-					middleware(req, res, next);
+				const middleware = createMiddleware(({ request }) => {
+					return miniflare.dispatchFetch(toMiniflareRequest(request), {
+						redirect: "manual",
+					}) as any;
 				});
-			};
+
+				handleWebSocket(
+					vitePreviewServer.httpServer,
+					miniflare.dispatchFetch,
+					vitePreviewServer.config.logger
+				);
+
+				return () => {
+					vitePreviewServer.middlewares.use((req, res, next) => {
+						middleware(req, res, next);
+					});
+				};
+			},
 		},
-		configurePreviewServer(vitePreviewServer) {
-			const miniflare = new Miniflare(
-				getPreviewMiniflareOptions(
-					vitePreviewServer,
-					pluginConfig.persistState ?? true
-				)
-			);
+		{
+			name: "vite-plugin-cloudflare:modules",
+			// We set `enforce: "pre"` so that this plugin runs before the Vite core plugins.
+			// Otherwise the `.wasm` extension cannot be used for module imports
+			enforce: "pre",
+			applyToEnvironment(environment) {
+				if (resolvedPluginConfig.type === "assets-only") {
+					return false;
+				}
 
-			const middleware = createMiddleware(({ request }) => {
-				return miniflare.dispatchFetch(toMiniflareRequest(request), {
-					redirect: "manual",
-				}) as any;
-			});
+				return Object.keys(resolvedPluginConfig.workers).includes(
+					environment.name
+				);
+			},
+			async resolveId(source, importer) {
+				if (!source.endsWith(".wasm")) {
+					return;
+				}
 
-			handleWebSocket(
-				vitePreviewServer.httpServer,
-				miniflare.dispatchFetch,
-				vitePreviewServer.config.logger
-			);
+				const resolved = await this.resolve(source, importer);
+				assert(
+					resolved,
+					`Unexpected error: could not resolve Wasm module ${source}`
+				);
 
-			return () => {
-				vitePreviewServer.middlewares.use((req, res, next) => {
-					middleware(req, res, next);
-				});
-			};
+				return {
+					external: true,
+					id: createModuleReference("CompiledWasm", resolved.id),
+				};
+			},
+			renderChunk(code, chunk) {
+				const moduleRE = new RegExp(MODULE_PATTERN, "g");
+				let match: RegExpExecArray | null;
+				let magicString: MagicString | undefined;
+
+				while ((match = moduleRE.exec(code))) {
+					magicString ||= new MagicString(code);
+					const [full, moduleType, modulePath] = match;
+
+					assert(
+						modulePath,
+						`Unexpected error: module path not found in reference ${full}.`
+					);
+
+					let source: Buffer;
+
+					try {
+						source = fs.readFileSync(modulePath);
+					} catch (error) {
+						throw new Error(
+							`Import ${modulePath} not found. Does the file exist?`
+						);
+					}
+
+					const referenceId = this.emitFile({
+						type: "asset",
+						name: path.basename(modulePath),
+						originalFileName: modulePath,
+						source,
+					});
+
+					const emittedFileName = this.getFileName(referenceId);
+					const relativePath = vite.normalizePath(
+						path.relative(path.dirname(chunk.fileName), emittedFileName)
+					);
+					const importPath = relativePath.startsWith(".")
+						? relativePath
+						: `./${relativePath}`;
+
+					magicString.update(
+						match.index,
+						match.index + full.length,
+						importPath
+					);
+				}
+
+				if (magicString) {
+					return {
+						code: magicString.toString(),
+						map: this.environment.config.build.sourcemap
+							? magicString.generateMap({ hires: "boundary" })
+							: null,
+					};
+				}
+			},
 		},
-	};
-
-	return [modulesPlugin(resolvedPluginConfigContainer), plugin];
+	];
 }
 
 /**
@@ -360,4 +452,8 @@ function getDotDevDotVarsContent(
 	}
 
 	return null;
+}
+
+function createModuleReference(type: ModuleType, id: string) {
+	return `__CLOUDFLARE_MODULE__${type}__${id}__`;
 }
